Split BasicTest.kt into several test files
