+++ src/test/kotlin/BasicTest.kt
@@ -34,59 +34,4 @@ class BasicTest {
             projectScope.getLocations() == desiredProjectScope
         )
     }
-
-    @Test
-    fun `Java Test Project Runs`() {
-        val testInputPath = "/testData/testProjects/javaTestProject"
-        val testSrcPath = "/testData/testProjects/javaTestProject/src/"
-        val exitCode = runIde(testInputPath, testSrcPath, ".")
-        assertEquals(0, exitCode, "The IDE should finish terminate with code 0")
-    }
-
-    @Test
-    fun `Dependency Between Two Files Detected - Java Test Project` () {
-        val testInputPath = "/testData/testProjects/javaTestProject"
-        val testSrcPath = "/testData/testProjects/javaTestProject/src/"
-        runIde(testInputPath, testSrcPath, ".")
-        Assert.assertFalse("Dependencies list is not empty", readTestFile() == "[]")
-        val dependencies = readFromJsonString(readTestFile())
-        Assert.assertTrue("Main.java should depend on Util.java",
-            dependencies.any { it.from.isInFile("Main.java") && it.to.isInFile("Util.java") })
-    }
-
-    @Test
-    fun `Dependency Within One File Detected - Java Test Project` () {
-        val testInputPath = "/testData/testProjects/javaTestProject"
-        val testSrcPath = "/testData/testProjects/javaTestProject/src/"
-        runIde(testInputPath, testSrcPath, ".")
-        Assert.assertFalse("Dependencies list is not empty", readTestFile() == "[]")
-        val dependencies = readFromJsonString(readTestFile())
-        val innerDependencies = dependencies.filter { it.from.location.path == it.to.location.path }
-        Assert.assertFalse("List has dependencies within one file", innerDependencies.isEmpty())
-    }
-
-    @Test
-    fun `Dependency Between Two Files Detected - Kotlin Test Project` () {
-        val testInputPath = "testData/testProjects/kotlinTestProject"
-        val testSrcPath = "/testData/testProjects/kotlinTestProject/src/"
-        val exitCode = runIde(testInputPath, testSrcPath, ".")
-        assertEquals(0, exitCode, "The IDE should finish terminate with code 0")
-        Assert.assertFalse("Dependencies list is not empty", readTestFile() == "[]")
-        println(readTestFile())
-        val dependencies = readFromJsonString(readTestFile())
-        Assert.assertTrue("Main.kt should depend on Utility.kt",
-                dependencies.any { it.from.isInFile("Main.kt") && it.to.isInFile("Utility.kt") })
-    }
-
-    @Test
-    fun `Dependency Within One File Detected - Kotlin Test Project` () {
-        val testInputPath = "testData/testProjects/kotlinTestProject"
-        val testSrcPath = "/testData/testProjects/kotlinTestProject/src/"
-        val exitCode = runIde(testInputPath, testSrcPath, ".")
-        assertEquals(0, exitCode, "The IDE should finish terminate with code 0")
-        Assert.assertFalse("Dependencies list is not empty", readTestFile() == "[]")
-        val dependencies = readFromJsonString(readTestFile())
-        val innerDependencies = dependencies.filter { it.from.location.path == it.to.location.path }
-        Assert.assertFalse("List has dependencies within one file", innerDependencies.isEmpty())
-    }
 }
